pyproject.toml
[tool.poetry]
name = "app-fastapi"
version = "0.1.0"
description = ""
authors = ["xmaksutx <xmaksutx@gmail.com>"]
readme = "README.md"
packages = [{include = "app_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"
mypy = "^1.5.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
